requirements.txt
flask
librosa
pydub
google-cloud-speech
mtcnn
facenet-pytorch
socketio
requests
pytest